move tariff initializer test to correct file
